+++ cirq-google/cirq_google/engine/engine_program_test.py
@@ -27,7 +27,7 @@ from cirq_google.engine import util
 from cirq_google.cloud import quantum
 from cirq_google.engine.engine import EngineContext
 from cirq_google.engine.result_type import ResultType
-from cirq_google.engine.util_test import uses_async_mock
+from cirq_google.engine.test_utils import uses_async_mock
 
 
 _BATCH_PROGRAM_V2 = util.pack_any(
