+++ cirq-google/cirq_google/engine/engine_test.py
@@ -29,7 +29,7 @@ from cirq_google.api import v1, v2
 from cirq_google.engine import util
 from cirq_google.cloud import quantum
 from cirq_google.engine.engine import EngineContext
-from cirq_google.engine.util_test import uses_async_mock
+from cirq_google.engine.test_utils import uses_async_mock
 
 
 _CIRCUIT = cirq.Circuit(
